pyproject.toml
[tool.poetry]
name = "Parsing_from_Wildberries"
version = "0.1.0"
description = ""
authors = ["Sergey Spisak <serega232003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
ipython = "^8.28.0"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
djangorestframework = "^3.16.0"
beautifulsoup4 = "^4.13.4"


[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
